+++ lib/sensors/dht22.h
@@ -2,7 +2,7 @@
 #define DHT_HEADER
 
 void dht_setup();
-void dht_record(double* temperature, double* humidity); // These will grab new readings from the DHT.
+void dht_record(float* temperature, float* humidity); // These will grab new readings from the DHT.
 void dht_print();
 
 #endif
